+++ cocos2d/CCMenu.m
@@ -122,7 +122,7 @@ enum {
 /*
  * override add:
  */
--(void) addChild:(CCMenuItem*)child z:(int)z tag:(int) aTag
+-(void) addChild:(CCMenuItem*)child z:(NSInteger)z tag:(NSInteger) aTag
 {
 	NSAssert( [child isKindOfClass:[CCMenuItem class]], @"Menu only supports MenuItem objects as children");
 	[super addChild:child z:z tag:aTag];
