+++ tests/SpriteTest.m
@@ -566,7 +566,7 @@ Class restartAction()
 {
 	id sprite = [self getChildByTag:kTagSprite1];
 	
-	int z = [sprite zOrder];
+	NSInteger z = [sprite zOrder];
 	
 	if( z < -1 )
 		dir = 1;
@@ -629,7 +629,7 @@ Class restartAction()
 	id batch = [self getChildByTag:kTagSpriteBatchNode];
 	id sprite = [batch getChildByTag:kTagSprite1];
 	
-	int z = [sprite zOrder];
+	NSInteger z = [sprite zOrder];
 	
 	if( z < -1 )
 		dir = 1;
@@ -673,21 +673,21 @@ Class restartAction()
 		//usually children get sorted before -transform but call sort now to verify order
 //		[asmtest sortAllChildren];
 		
-		int prev = -1;
+		NSInteger prev = -1;
 		for(id child in asmtest.children)
 		{
-			int currentIndex = [child atlasIndex];
+			NSUInteger currentIndex = [child atlasIndex];
 			NSAssert( prev == currentIndex-1, @"Child order failed");
-			NSLog(@"children %x - atlasIndex:%d", (NSUInteger)child, currentIndex);
+			NSLog(@"children %x - atlasIndex:%d", (unsigned int)child, (unsigned int) currentIndex);
 			prev = currentIndex;
 		}
 		
 		prev = -1;
 		for(id child in asmtest.descendants)
 		{
-			int currentIndex = [child atlasIndex];
+			NSUInteger currentIndex = [child atlasIndex];
 			NSAssert( prev == currentIndex-1, @"Child order failed");
-			NSLog(@"descendant %x - atlasIndex:%d", (NSUInteger)child, currentIndex);
+			NSLog(@"descendant %x - atlasIndex:%d", (unsigned int)child, (unsigned int) currentIndex);
 			prev = currentIndex;
 		}		
 	}
